Basic Graph Plotter
This is a very simple C# wpf application for plotting graph of different
mathmatical equation. Just type the function add it to the list and then
draw its corresponding graph.
To parse the mathmatical equation Math OS parser is used. This can be
downloaded from here (https://mathosparser.codeplex.com/)